requirements.txt
aniso8601==4.0.1
apispec==0.39.0
bcrypt==3.1.4
boto3==1.7.62
botocore==1.10.62
cffi==1.11.5
click==6.7
docutils==0.14
flasgger==0.9.2
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Login==0.4.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10.1
jmespath==0.9.3
jsonschema==2.6.0
MarkupSafe==1.0
mistune==0.8.4
numpy==1.16.0
Pillow==5.2.0
psycopg2==2.7.5
pycparser==2.18
python-dateutil==2.7.3
pytz==2018.7
PyYAML==3.13
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.3.3
-e git+https://github.com/temporaryorgname/database.git@de6aa0e33629a79b587bb4f82f7e3022e9ea5cfc#egg=tracker_database
Werkzeug==0.14.1
